Changing & moving sensitive data to config.ts

+++ src/card/card.service.ts
@@ -1,4 +1,4 @@
-import { HttpException, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserCredentials } from 'src/authentication/user.credentials';
 import { ColumnEntity } from 'src/column/column.entity';
@@ -22,7 +22,7 @@ export class CardService {
     private async columnExistenceCheck(columnId: number) {
         const column = await this.columnRepository.findOne(columnId);
         if (!column) 
-            throw new HttpException({ Column: ' not found' }, 404);
+            throw new HttpException({ Column: ' not found' }, HttpStatus.NOT_FOUND);
     }
 
     async create(userCreds: UserCredentials, columnId: number, data: CardDtoRequest): Promise<CardDtoResponse> {
@@ -35,7 +35,7 @@ export class CardService {
             relations: ['cards']
         })
         if (!user)
-            throw new HttpException({ CardUser: ' not found'}, 404)
+            throw new HttpException({ CardUser: ' not found'}, HttpStatus.NOT_FOUND)
 
         const column = await this.columnRepository.findOne({
             where: {
@@ -67,7 +67,7 @@ export class CardService {
             relations: ['user']
         })
         if (!card)
-            throw new HttpException({ Card: ' not found'}, 404)
+            throw new HttpException({ Card: ' not found'}, HttpStatus.NOT_FOUND)
 
         return new CardDtoResponse(card, card.user.id, columnId )
     }
@@ -87,7 +87,7 @@ export class CardService {
             new CardDtoResponse(card, card.user.id, columnId))
     }
 
-    async update(columnId: number, cardId: number, data: CardDtoRequest): Promise<CardDtoResponse> {
+    async update(userCreds: UserCredentials, columnId: number, cardId: number, data: CardDtoRequest): Promise<CardDtoResponse> {
         await this.columnExistenceCheck(columnId);
         const card = await this.cardRepository.findOne({
             where: {
@@ -96,19 +96,27 @@ export class CardService {
                     id: columnId
                 }
             },
-            relations: ['user']
+            relations: ['user', 'column']
         })
         if (!card)
-            throw new HttpException({ Card: ' not found'}, 404)
+            throw new HttpException({ Card: ' not found'}, HttpStatus.NOT_FOUND)
 
+        if (card.user.id != userCreds.userId)
+            throw new HttpException('no permission', HttpStatus.UNAUTHORIZED)
+        
+        if (columnId != data.columnId) {
+            await this.columnExistenceCheck(data.columnId)
+            const newColumn = await this.columnRepository.findOne(data.columnId) 
+            card.column = newColumn;
+        }
         card.title = data.title;
         card.description = data.description;
         this.cardRepository.save(card);
 
-        return new CardDtoResponse(card, card.user.id, columnId);
+        return new CardDtoResponse(card, card.user.id, card.column.id);
     }
 
-    async delete(columnId: number, cardId: number) {
+    async delete(userCreds: UserCredentials, columnId: number, cardId: number) {
         await this.columnExistenceCheck(columnId);
         const card = await this.cardRepository.findOne({
             where: {
@@ -116,10 +124,13 @@ export class CardService {
                 column: {
                     id: columnId
                 }
-            }
+            }, relations: ['user']
         })
         if (!card)
-            throw new HttpException({ Card: ' not found'}, 404)
+            throw new HttpException({ Card: ' not found'}, HttpStatus.NOT_FOUND)
+
+        if (card.user.id != userCreds.userId)
+            throw new HttpException('no permission', HttpStatus.UNAUTHORIZED)
 
         return await this.cardRepository.delete(cardId)
     }
