+++ src/column/column.service.ts
@@ -1,4 +1,4 @@
-import { HttpException, Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserCredentials } from 'src/authentication/user.credentials';
 import { UserEntity } from 'src/user/user.entity';
@@ -27,7 +27,7 @@ export class ColumnService {
         });
 
         if (!user) 
-            throw new HttpException({ User: ' not found' }, 404);
+            throw new HttpException({ User: ' not found' }, HttpStatus.NOT_FOUND);
 
         user.columns.push(сolumn);
         сolumn = await this.columnRepository.save(сolumn);
@@ -44,7 +44,7 @@ export class ColumnService {
         });
 
         if (!user) 
-            throw new HttpException({ User: ' not found' }, 404);
+            throw new HttpException({ User: ' not found' }, HttpStatus.NOT_FOUND);
 
         return user.columns.map(element => 
             new ColumnDtoResponse(element.id, element.name, user.id, element.createdAt)
@@ -54,7 +54,7 @@ export class ColumnService {
     async getById(id: number, userId: number): Promise<ColumnDtoResponse> {
         const user = await this.userRepository.findOne(userId);
         if (!user) 
-            throw new HttpException({ User: ' not found' }, 404);
+            throw new HttpException({ User: ' not found' }, HttpStatus.NOT_FOUND);
 
         const column = await this.columnRepository.findOne({ 
             where: { 
@@ -67,47 +67,53 @@ export class ColumnService {
         });
 
         if (!column) 
-            throw new HttpException({ User: ' invalid' }, 404)
+            throw new HttpException({ User: ' invalid' }, HttpStatus.NOT_FOUND)
         
         return new ColumnDtoResponse(column.id, column.name, user.id, column.createdAt)
     }
 
-    async update(id: number, userData: UserCredentials, newData: ColumnDtoRequest) {
-        const user = await this.userRepository.findOne(userData.userId);
+    async update(id: number, userCreds: UserCredentials, newData: ColumnDtoRequest) {
+        const user = await this.userRepository.findOne(userCreds.userId);
         if (!user) 
-            throw new HttpException({ User: ' not found' }, 404);
+            throw new HttpException({ User: ' not found' }, HttpStatus.NOT_FOUND);
         
         const column = await this.columnRepository.findOne({ 
             where: { 
                 id: id,
                 user: {
-                    id: userData.userId
+                    id: userCreds.userId
                 }
-            } 
+            }, relations: ['user']
         });
         if (!column) 
-            throw new HttpException({ Column: ' not found' }, 404);
+            throw new HttpException({ Column: ' not found' }, HttpStatus.NOT_FOUND);
+
+        if (column.user.id != userCreds.userId)
+            throw new HttpException('no permission', HttpStatus.UNAUTHORIZED)
 
         column.name = newData.name;
         return await this.columnRepository.save(column)
     }
 
-    async delete(id: number, userData: UserCredentials) {
-        const user = await this.userRepository.findOne(userData.userId);
+    async delete(id: number, userCreds: UserCredentials) {
+        const user = await this.userRepository.findOne(userCreds.userId);
         if (!user) 
-            throw new HttpException({ User: ' not found' }, 404);
+            throw new HttpException({ User: ' not found' }, HttpStatus.NOT_FOUND);
         
         const column = await this.columnRepository.findOne({ 
             where: { 
                 id: id,
                 user: {
-                    id: userData.userId
+                    id: userCreds.userId
                 }
-            } 
+            }, relations: ['user']
         });
         if (!column) 
-            throw new HttpException({ Column: ' not found' }, 404);
-            
+            throw new HttpException({ Column: ' not found' }, HttpStatus.NOT_FOUND);
+        
+        if (column.user.id != userCreds.userId)
+            throw new HttpException('no permission', HttpStatus.UNAUTHORIZED)
+
         return await this.columnRepository.delete(id)
     }
 }
